refactor: rename api file into its hook's name
